requirements.txt
# Core dependencies
tensorflow>=2.12.0
opencv-python>=4.7.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
Pillow>=9.4.0

# Web framework
flask>=2.2.0
flask-cors>=3.0.10
gunicorn>=20.1.0

# API
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0

# Hardware support
picamera2>=0.3.9  # For Raspberry Pi
pyserial>=3.5  # For Arduino communication

# Utils
python-dotenv>=0.21.0
requests>=2.28.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.2.0
black>=23.1.0
flake8>=6.0.0
pre-commit>=3.0.0
